requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
numpy>=1.24.2
Pillow>=9.4.0
torch>=2.0.0
torchvision>=0.15.1
opencv-python>=4.7.0
shapely>=2.0.1
scipy>=1.10.1
matplotlib>=3.7.1
torch>=2.0.0
torchvision>=0.15.1
opencv-python>=4.7.0
shapely>=2.0.1
scipy>=1.10.1
matplotlib>=3.7.1
